+++ src/LDL/Framework/Base/Exception/ToArrayException.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace LDL\Framework\Base\Exception;
-
-class ToArrayException extends LDLException
-{
-
-}