+++ CHANGELOG.md
@@ -42,6 +42,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - feature/1201243443283062 - Add __toString in ToStringInterface / Pass original items to IterableHelper::map
 - feature/1201253503567912 - Add constants to type interfaces
 - feature/1201278499056880 - Add $preserveKeys argument to IterableHelper::map (true by default)
+- feature/1201329585503220 - IterableHelper::toArray takes in account ToArrayInterface
 
 ### Changed